Fix horizontal bar charts not updating correctly (#54)
Horizontal bar charts didn't update correctly - the data that the width
was calculated by was bound to previous values.

The fix was to use `select()` instead of `selectAll` on the bars
themselves (since there is only a single `rect` element in each). This
is because `selectAll` doesn't propagate data bindings, per the
following links:

https://groups.google.com/forum/?fromgroups#!topic/d3-js/PUansS150J0
https://groups.google.com/forum/?fromgroups#!topic/d3-js/272vChmijIw

It works the first time because of the `.enter()` definition which will
cause data to propagate, as mentioned in the above threads.