Fix a bug where the tooltip might show an incorrect stream (#62)
If you disabled a series in a multi-bar chart (either horizontal or
vertical), you would get incorrect tooltips. This was caused by the
fact that we only set the mouseover on the `enter()` elements, which
meant that nothing was being set when the chart was updated due to the
hiding. This caused us to close over the original value of the data

The fix is to set the mouseover on the entire selection, which updates
it every time the chart is updated. This also contains a small fix to
the horizontal multi-bar chart to have a more descriptive tooltip.