left, bottom and right label positioning (+ some bugs)
The ability to position the left label text is there through axisLabelDistance, however the right and bottom labels for a chart do not have this functionality. It has now been added (used the same way for all three)

xAxis.axisLabelDistance...
yAxis.axisLabelDistance...
y1Axis.axisLabelDistance...
y2Axis.axisLabelDistance...

I also came a cross a few errors while dealing with multiple charts on a page where sometimes (not all the time) there would be an error on mouse over of sections of a chart. This was due to the d3.event being undefined and then later it using d3.event.pageX or d3.event.pageY. I threw in checks to see if d3.event was truethy and if so continue on, if not, do nothing.