Fixes custom x/y value getters for multiChart
The multiChart model provides `x()` and `y()` methods like many other nv models, even with the same implementation. But there are some default settings missing, which causes the methods to throw an error: `Uncaught ReferenceError: getX is not defined`. I added the two missing variable declarations with a default method based on the conventions in other models.