Use nvd3-specific CSS rules
Simply using `svg` is not a good idea as it can interfere with other `svg` elements that do not use `nvd3` but just pure `d3`. Changing `svg` to `nvd3svg` means that every `svg` element that will be used by `nvd3` should be given an `nvd3svg` class for these CSS rules to take effect.