Only emit svg dom elements for active points
If a given point is marked as not active, e.g.

    { 'x': 1, 'y': 5, 'notActive': true }

Then when a scatter plot is drawn (not just with scatterChart(), but
also for any other chart that uses models.scatter, like lineChart(),
stackAreaChart(), etc.), that point won't show up via mouseovers or the
interactive guide, but the svg dom elements for that point are still
emitted. When one has thousands of points to display, but we only want
to highlight a few of them, or perhaps just draw the points with no
iteractions at all, we still pay a performance penalty for those
elements.

As of this commit if a point is marked 'notActive' then no svg dom
elements are emitted to draw it.  While this does not make sense to use
this behavior for a scatterChart(), it is very helpful for lineChart()s
and stackedAreaChart()s (and others) which include scatter charts for
the points.