Restoring default color behaviour for legends
Now the default behaviour (without explicitly calling chart.color() or
chart.barColor() is shades of gray, a different shade per series. This
means that when chart.barColor() behaviour is triggered, the legend
will stay in shades of grey to match the different shading of the
series. Moreover, unlike before, the per series shades of gray
colouring can be overridden by calling chart.color() if colourful per
series colouring is desired.