Move minimum D3 version to 3.4.4
As of D3 commit:

https://github.com/mbostock/d3/commit/3c7cc81b6233289263525b581d635ad9f8d9dd37

We are no longer able to use an array accessor that accepts two
parameters (the second being the index of the first parameter in the
given array), since d3.bisector() will consider it a comparison
function instead of an accessor.

There does not appear to be a way to get the index value of the element
being considered in the context of the accessor.

As a result, this commit proposes we drop the interactiveBisect()
support for accessors accepting a second parameter being the index of
the first in a given array.

We now wrap all x-accessors passed to interactiveBisect() hiding the
second parameter, using an internal comparator instead.

This commit also attempts to get the coverage to 100% for
interactiveBisect().

See issue #754 as well.