Using voronoi, collapse siblings, display all
When using voronoi points for tooltips on any chart relying on a scatter
chart underneath, collapse points to unique values that D3 accepts
(basically, points are considered duplicates with 1e-6 of each other).
But instead of hiding the values considered duplicates, show them all in
the same toolip.

FIXME: this is only implemented for lineChart, still need to amend this
commit with changes to the others that use scatter underneath.