Added maxKeyLength variable to truncate legend key values.
This fix provides an additional variable to set a fixed max key length
value to truncate any legend key values with lengths over the specified
threshold. This provides the option to organize legend keys to be seen
in a more readable manner.

Added tooltip for truncated legend text.
When mousing over truncated text within the legend,
a tooltip showing the full key value is revealed.