Add ability to change focus graph margins
For both line-with-focus-chart and line-plus-bar-chart, if one would
want to place an x-axis label below both the graph and its focus graph,
prior to this patch it is not possible in a fixed height setting, as
there is no way to change the bottom margin to account for a label.

This change adds the ability to change the bottom margin.  Note that in
order for that label to work while maintaining the focus graph height,
one also has to up the "focusHeight" value.