Turn off scatter pts in focus charts
A line-with-focus-chart or a line-plus-bar-chart will essentially
duplicate the number of lines and points drawn on the screen because
there are two copies of an nv.models.line maintained.  We only need
the scatter points for the main chart, the focus chart will never
use them.  This can save a lot of memory and CPU when rendering
graphs with large data sets.