Bugfix in initTooltip() function
When there are several charts, they share the same tooltip! In some exceptional cases, DOM element is removed but NDV3 tooltip object still exist! This causes initTooltip function to not create new tooltip. As the result, NVD3 tooltip object will be updated but since it doesn't exist in DOM elements it won't be shown anywehre.