Improve Sunburst margin & show percentage on tooltip
This commit fixes the margin on the Sunburst chart.
Since the sunburstChart.js has no longer a wrapper container,
the margin options are useless. Instead, the margin is now
shared between the sunburstChart and the sunburst objects.
Furthermore, the sunburst.js did not take into account the
margin when calculating the middle of the container.
Also, when the update() function was called the container
did not apply the newly calculated translate() arguments.

Finally, the sunburstChart.js now has the option to show
the percentage of a node in the tooltip, similar to the
pieChart.