Fix lineWithFocusChart rendering issues
This commit fixes a rendering issue wit the lineWithFocusChart.
When a lineWithFocusChart initially has no data, then gets data and
update() is called, only the focus part is rendered. Only when data is
set again and update() is called again does the 'normal' linechart
show correctly.
Listening for the 'onBrush' event of the focus before initializing the
chart seems to fix the issue.