fix(sunburst): pass more data to color method
This fix passes the entire data object and index into the color function rather than just the name. In addition to to allowing more intelligent color functions, this change makes the API more consistent with the pie chart.

BREAKING CHANGE:

Any existing sunburst chart that uses a custom color function will need to be updated to the new api; however, the name is still passed to this function as an attribute of the object.