fix: Line chart interactive guideline error when extent is not in ascending order
Fixes a “TypeError: Cannot read property 'x' of undefined” error when
using the interactive guideline for line charts if the extent is not in
ascending order.

In our use case, we reversed the x-axis on the line chart by reversing
the x domain.  This caused series.values.filter to filter out all our
values and result in the TypeError when calling interactiveBisect  A
check was added to ensure the extent is in ascending order.