Don't render focus until shown for first time
Several unit tests assumed that the focus chart would not exist in the DOM initially, and it makes sense to maintain this optimization as it might never be enabled.  Prior to first rendering, focus.brush.empty() returns false so we can't rely on that, but focus.brush.extent() will be null.